Add WorkspaceDocument+SearchState and WorkspaceDocument+Index
